feat: #185 경매 엔티티에 판매자 필드 추가 (#187)
* feat: 경매 엔티티에 회원 엔티티 연관관계 추가

* feat: 경매 등록 시 판매자를 등록하는 로직 추가

* feat: 경매 조회 시 판매자 fetch join 기능 추가

* feat: 경매 등록 및 조회 api 판매자 연동

* feat: 경매 하위 카테고리 연관관계 세팅 기능 추가

* refactor: 하위 카테고리를 생성자로 초기화하지 않고 연관관계 세팅 메서드를 통해 초기화하도록 변경

* refactor: 판매자, 하위 카테고리 연관관계 세팅 로직 메서드 분리

* chore: 경매 엔티티 변경에 따른 flyway 스크립트 추가

* test: 경매 엔티티 관련 컴파일 에러 해결

* test: 테스트 어노테이션 변경

* feat: 특정 회원과 경매 판매자가 일치하는지 여부를 판단하는 기능 추가

* feat: 경매 삭제 시 판매자임을 검증하는 기능 추가

+++ backend/ddang/src/main/java/com/ddang/ddang/auction/application/AuctionService.java
@@ -5,8 +5,10 @@ import com.ddang.ddang.auction.application.dto.CreateInfoAuctionDto;
 import com.ddang.ddang.auction.application.dto.ReadAuctionDto;
 import com.ddang.ddang.auction.application.dto.ReadAuctionsDto;
 import com.ddang.ddang.auction.application.exception.AuctionNotFoundException;
+import com.ddang.ddang.auction.application.exception.UserNotAuthorizationException;
 import com.ddang.ddang.auction.domain.Auction;
 import com.ddang.ddang.auction.infrastructure.persistence.JpaAuctionRepository;
+import com.ddang.ddang.bid.application.exception.UserNotFoundException;
 import com.ddang.ddang.category.application.exception.CategoryNotFoundException;
 import com.ddang.ddang.category.domain.Category;
 import com.ddang.ddang.category.infrastructure.persistence.JpaCategoryRepository;
@@ -17,6 +19,8 @@ import com.ddang.ddang.region.application.exception.RegionNotFoundException;
 import com.ddang.ddang.region.domain.AuctionRegion;
 import com.ddang.ddang.region.domain.Region;
 import com.ddang.ddang.region.infrastructure.persistence.JpaRegionRepository;
+import com.ddang.ddang.user.domain.User;
+import com.ddang.ddang.user.infrastructure.persistence.JpaUserRepository;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -29,6 +33,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AuctionService {
 
+    private final JpaUserRepository userRepository;
     private final JpaAuctionRepository auctionRepository;
     private final JpaRegionRepository regionRepository;
     private final JpaCategoryRepository categoryRepository;
@@ -36,25 +41,32 @@ public class AuctionService {
 
     @Transactional
     public CreateInfoAuctionDto create(final CreateAuctionDto dto) {
-        final Auction auction = convertAuction(dto);
+        final Auction auction = dto.toEntity();
+        final User seller = findSeller(dto);
+        final Category subCategory = findSubCategory(dto);
         final List<AuctionRegion> auctionRegions = convertAuctionRegions(dto);
         final List<AuctionImage> auctionImages = convertAuctionImages(dto);
 
         auction.addAuctionRegions(auctionRegions);
         auction.addAuctionImages(auctionImages);
+        auction.addSeller(seller);
+        auction.addSubCategory(subCategory);
 
         final Auction persistAuction = auctionRepository.save(auction);
 
         return CreateInfoAuctionDto.from(persistAuction);
     }
 
-    private Auction convertAuction(final CreateAuctionDto dto) {
-        final Category subCategory = categoryRepository.findSubCategoryById(dto.subCategoryId())
-                                                       .orElseThrow(() -> new CategoryNotFoundException(
-                                                               "지정한 하위 카테고리가 없거나 하위 카테고리가 아닙니다."
-                                                       ));
+    private User findSeller(final CreateAuctionDto dto) {
+        return userRepository.findById(dto.sellerId())
+                             .orElseThrow(() -> new UserNotFoundException("지정한 판매자를 찾을 수 없습니다."));
+    }
 
-        return dto.toEntity(subCategory);
+    private Category findSubCategory(final CreateAuctionDto dto) {
+        return categoryRepository.findSubCategoryById(dto.subCategoryId())
+                                 .orElseThrow(() -> new CategoryNotFoundException(
+                                         "지정한 하위 카테고리가 없거나 하위 카테고리가 아닙니다."
+                                 ));
     }
 
     private List<AuctionRegion> convertAuctionRegions(final CreateAuctionDto dto) {
@@ -79,12 +91,16 @@ public class AuctionService {
     }
 
     public ReadAuctionDto readByAuctionId(final Long auctionId) {
-        final Auction auction = auctionRepository.findAuctionById(auctionId)
-                                                 .orElseThrow(() -> new AuctionNotFoundException(
-                                                         "지정한 아이디에 대한 경매를 찾을 수 없습니다."
-                                                 ));
+        final Auction auction = findAuction(auctionId);
 
         return ReadAuctionDto.from(auction);
+    }
+
+    private Auction findAuction(final Long auctionId) {
+        return auctionRepository.findAuctionById(auctionId)
+                                .orElseThrow(() -> new AuctionNotFoundException(
+                                        "지정한 아이디에 대한 경매를 찾을 수 없습니다."
+                                ));
     }
 
     public ReadAuctionsDto readAllByLastAuctionId(final Long lastAuctionId, final int size) {
@@ -94,12 +110,19 @@ public class AuctionService {
     }
 
     @Transactional
-    public void deleteByAuctionId(final Long auctionId) {
-        final Auction auction = auctionRepository.findById(auctionId)
-                                                 .orElseThrow(() -> new AuctionNotFoundException(
-                                                         "지정한 아이디에 대한 경매를 찾을 수 없습니다."
-                                                 ));
+    public void deleteByAuctionId(final Long auctionId, final Long userId) {
+        final Auction auction = findAuction(auctionId);
+        final User user = findUser(userId);
+
+        if (!auction.isOwner(user)) {
+            throw new UserNotAuthorizationException("권한이 없습니다.");
+        }
 
         auction.delete();
+    }
+
+    private User findUser(final Long userId) {
+        return userRepository.findById(userId)
+                                        .orElseThrow(() -> new UserNotFoundException("회원 정보를 찾을 수 없습니다."));
     }
 }
